requirements.txt
fastapi~=0.116.1
sqlalchemy~=2.0.43
pydantic-settings~=2.10.1
python-jose~=3.5.0
passlib~=1.7.4
pydantic~=2.11.9
python-multipart~=0.0.20
pydantic-settings~=2.10.1
psycopg2-binary~=2.9.10
email-validator~=2.3.0
cryptography~=45.0.7
bcrypt~=4.3.0
uvicorn~=0.35.0







